Undo hjkl arrow keys
I noticed that my hand position on the home row has been completely
wrong ALL. THOSE. YEARS. And I blame vim. When I started learning vim I
used hjkl a lot which meant that it made sense for my right hand to rest
in the hjkl position. That shifted my left hand one key to the left,
which is not  the proper home row position. With the planck this end up
being extremely annoying because my thumbs are supposed to rest more or
less on the space keys when in fact the left hand sits on the lower key.
That causes so much friction. I hope that by undoing this change and
disabling hjkl in vim I can teach myself the proper asdf jkl; position